requirements.txt
transformers
huggingface_hub
python-dotenv
torch 
torchvision 
torchaudio 
numpy<2